+++ pages/examples/products/[id].js
@@ -60,7 +60,7 @@ export default function Page ({ app, product }) {
           <div className="mt-10 lg:mt-0 lg:col-start-2 lg:row-span-2 lg:self-center">
             <div className="aspect-w-1 aspect-h-1 rounded-lg overflow-hidden">
               <img
-                src={`${app.url}/assets/${product.thumbnail}`}
+                src={`${app.apiUrl}/assets/${product.thumbnail}`}
                 alt={product.name}
                 className="w-full h-full object-center object-cover"
               />